requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
mysql==0.0.2
mysqlclient==1.4.4
PyMySQL==0.9.3
python-dotenv==0.10.3
Werkzeug==0.16.0
